+++ udata_front/theme/gouvfr/assets/js/components/AdminSidebar/AdminSidebarMenu/AdminSidebarMenu.vue
@@ -64,11 +64,19 @@
             icon="fr-icon-parent-line"
             :label="t('Members')"
             :to="{name: 'organization-members', params: {oid: organization.id}}"
+            @click="$emit('click')"
+          />
+          <AdminSidebarLink
+            icon="fr-icon-server-line"
+            :label="t('Harvesters')"
+            :to="{name: 'organization-harvesters', params: {oid: organization.id}}"
+            @click="$emit('click')"
           />
           <AdminSidebarLink
           icon="fr-icon-user-line"
           :label="t('Profile')"
           :to="{name: 'organization-profile', params: {oid: organization.id}}"
+          @click="$emit('click')"
         />
         </template>
       </ul>