+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Add special style to tabular API dataservice card [#548](https://github.com/datagouv/udata-front/pull/548)
+- Export `Placeholder` component, use reuse last_modified [#546](https://github.com/datagouv/udata-front/pull/546)
 
 ## 1.4.0 (2024-09-23)